refactor(elasticsearch): update ES_INDICES import paths and enhance type safety in ElasticsearchService methods
